requirements.txt
crewai==0.28.8
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
